stream_processing/producer/requirements.txt
kafka-python==2.0.2
hdfs==2.6.0